[Feature] Validating HTTP Method to Detele Function

+++ App/Services/UserService.php
@@ -41,12 +41,18 @@ class UserService {
         if ($method != "PUT") {
             return throw new \Exception("HTTP method must be 'PUT'");
         }
-        
+
         $_PUT = file_get_contents('php://input');
         return User::update(json_decode($_PUT));
     }
 
     public function delete() {
+        $method = $_SERVER['REQUEST_METHOD'];
+
+        if ($method != "DELETE") {
+            return throw new \Exception("HTTP method must be 'DELETE'");
+        }
+        
         $_DELETE = file_get_contents('php://input');
         return User::delete(json_decode($_DELETE));
     }
